requirements.txt
AutoDotEnv==0.1.3
click==8.1.3
pyotp==2.6.0
python-dotenv==0.20.0
urllib3==1.26.9
